Initial Commit
This cl creates the overall bulk of the site. We have a canvas
that renders a start and end point, and as the user interacts
with the canvas (mousedown) we add/update the control points.

Using this a user can get a better idea how to draw a bezier
curve, along with the different ways that it renders.